Improve {lib,}jvm.{so,dylib,dll} loading
Use the values from  Makeconf to find java (Unix), falling back to
JAVA_HOME (which is needed on Windows). Add the directory to PATH
on Windows or dyn.load the shared library it on Linux. macOS is
kept largely the same as it was except we check Makeconf before
JAVA_HOME.